Move gl account mapping seed
